Migrate improvements from boxcutter repo

+++ bin/lint.sh
@@ -3,10 +3,18 @@
 set -eu
 set -o pipefail
 
+HADOLINT_CONTAINER_IMAGE=polymathrobotics/hadolint:2.10.0
+
 SCRIPT_PATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 BIN_DIR="${SCRIPT_PATH}"
-BASE_DIR=$(pwd)
-HADOLINT_CONTAINER_IMAGE=polymathrobotics/hadolint
+CONTAINERFILE_DIR=$(pwd)
 
-lints_to_ignore=$(sed "s/#.*//" "${BIN_DIR}/hadolint-ignore" | sed '/^[[:space:]]*$/d' | sed 's/^/--ignore / ' | tr '\n' ' ')
-docker container run --rm -i ${HADOLINT_CONTAINER_IMAGE} hadolint ${lints_to_ignore} - < "Containerfile"
+"${BIN_DIR}/check-image.sh" "${HADOLINT_CONTAINER_IMAGE}"
+
+if [[ -f "${BIN_DIR}/hadolint-ignore" ]]; then
+  lints_to_ignore=$(sed "s/#.*//" "${BIN_DIR}/hadolint-ignore" | sed '/^[[:space:]]*$/d' | sed 's/^/--ignore / ' | tr '\n' ' ')
+  docker container run --rm -i ${HADOLINT_CONTAINER_IMAGE} hadolint ${lints_to_ignore} - < "Containerfile"
+  exit 0
+fi
+
+docker container run --rm -i ${HADOLINT_CONTAINER_IMAGE} hadolint - < "Containerfile"
